backend_python/requirements-core.txt
# FastAPI and dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.7
alembic==1.12.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Environment & Configuration
python-dotenv==1.0.0
pydantic-settings==2.0.3

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2